requirements.txt
torch
transformers
Pillow